// app/src/bp-api/models/FundingIntervalRate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FundingIntervalRate = {
  /**
   * The symbol of the market associated to the funding interval.
   */
  symbol: string
  /**
   * The end of the funding interval.
   */
  intervalEndTimestamp: string
  /**
   * The funding rate for the interval.
   */
  fundingRate: string
}