// app/src/bp-api/models/OpenInterest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OpenInterest = {
  /**
   * The symbol of the market.
   */
  symbol: string
  /**
   * The open interest.
   */
  openInterest?: string
  /**
   * Timestamp.
   */
  timestamp: number
}