// app/src/bp-api/models/Ticker.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Ticker = {
  /**
   * The symbol of the market.
   */
  symbol: string
  /**
   * First price for the interval.
   */
  firstPrice: string
  /**
   * Last price for the interval.
   */
  lastPrice: string
  /**
   * Price change for the interval.
   */
  priceChange: string
  /**
   * Percentage price change for the interval.
   */
  priceChangePercent: string
  /**
   * Highest price for the interval.
   */
  high: string
  /**
   * Lowest price for the interval.
   */
  low: string
  /**
   * Base volume for the interval.
   */
  volume: string
  /**
   * Quote volume for the interval.
   */
  quoteVolume: string
  /**
   * Trades for the interval.
   */
  trades: string
}